[FE]Rename: 유저 관련 폴더명 및 폴더구조 변경
